+++ genome/gff.py
@@ -2,7 +2,7 @@
 """
  * @Date: 2022-10-12 19:32:50
  * @LastEditors: hwrn hwrn.aou@sjtu.edu.cn
- * @LastEditTime: 2025-01-13 21:07:59
+ * @LastEditTime: 2025-02-07 11:38:05
  * @FilePath: /genome/genome/gff.py
  * @Description:
 """
@@ -11,10 +11,12 @@ import gzip
 import warnings
 from pathlib import Path
 from typing import Callable, Generator, Iterable, TextIO
+import re
 
 import gffutils
 import gffutils.feature
 from Bio import SeqIO
+from Bio.Data import IUPACData
 from Bio.SeqFeature import SeqFeature, SimpleLocation
 from Bio.SeqRecord import SeqRecord
 from gffutils.exceptions import EmptyInputError
@@ -146,7 +148,7 @@ class InferGeneId:
 def infer_gene_id(rec_id: str | None, fet: SeqFeature):
     if fet.qualifiers and (name := fet.qualifiers.get("Name")) and name[0]:
         return name[0]
-    return f"{rec_id}_" + str(int(fet.id.rsplit("_", 1)[1]))
+    return f"{rec_id}_" + str(int(fet.id.rsplit("_", 1)[-1]))
 
 
 @InferGeneId.rec
@@ -200,6 +202,88 @@ def to_seqfeature(feature: gffutils.feature.Feature):
         type=feature.featuretype,
         qualifiers=qualifiers,
     )
+
+
+# region TranslExcept
+class TranslExcept:
+    PATTERN = re.compile(
+        r"\(pos:([0-9]+)\.\.([0-9]+),aa:([A-Za-z]+)\)|"
+        r"\(pos:complement\(([0-9]+)\.\.([0-9]+)\),aa:([A-Za-z]+)\)"
+    )
+
+    def __new__(cls, text: "str|TranslExcept"):
+        if isinstance(text, cls):
+            return text
+        return super().__new__(cls)
+
+    def __init__(self, text: str):
+        # GFF entry 是一行 GFF 文件的记录 (字符串)
+        self.text = text
+        _region, self.strand = self.parse_transl_except(text)
+        self.start = int(_region[0])
+        self.end = int(_region[1])
+        self.aa: str = _region[2]
+
+    @classmethod
+    def parse_transl_except(cls, text: str):
+        """
+        AP024703.1	DDBJ	CDS	3092001	3095066	.	+	0	ID=cds-BCX53216.1;Note=codon on position 197 is selenocysteine opal codon.;transl_except=(pos:3092589..3092591,aa:Sec)
+        JACSQK010000001.1	Protein Homology	CDS	324590	327646	.	-	0	ID=cds-MBD7959139.1;transl_except=(pos:complement(327056..327058),aa:Sec)
+        """
+        matches = cls.PATTERN.findall(text)
+        if len(matches) != 1:
+            raise ValueError(f"Invalid transl_except: {text}=>{matches}")
+        if matches[0][:3] == ("", "", ""):
+            return matches[0][3:], -1
+        if matches[0][3:] == ("", "", ""):
+            return matches[0][:3], 1
+        raise NotImplementedError(f"{matches[0][:3]} {matches[0][3:]}")
+
+    def index(self, position: SimpleLocation):
+        """
+        >>> TranslExcept("(pos:3092589..3092591,aa:Sec)").index(SimpleLocation(3092001, 3095066, 1))
+        (196, 'Sec')
+        >>> TranslExcept("(pos:complement(327056..327058),aa:Sec)").index(SimpleLocation(324590, 327646, -1))
+        (196, 'Sec')
+        """
+        assert self.strand == position.strand
+        if self.strand == 1:
+            return (self.start - position.start) // 3, self.aa
+        elif self.strand == -1:
+            return (position.end - self.end) // 3, self.aa
+        raise NotImplementedError
+
+    @classmethod
+    def to_str(cls, location: SimpleLocation, *transl_except: str):
+        return ";".join(
+            ("{0}@{1}".format(*cls(i).index(location)) for i in transl_except)
+        )
+
+    @classmethod
+    def un_str(cls, text: str | SeqRecord):
+        if isinstance(text, SeqRecord):
+            text = str(text.annotations.get("transl_except", ""))
+        if not text:
+            return
+        for i in text.split(";"):
+            i, a = i.split("@")
+            yield int(i), a
+
+    @classmethod
+    def modify(cls, aa, text: str, how="X"):
+        """
+        U = "Sec";  selenocysteine
+        O = "Pyl";  pyrrolysine
+        """
+        for j, ea in cls.un_str(text):
+            if ea in IUPACData.protein_letters_3to1_extended:
+                ea = IUPACData.protein_letters_3to1_extended[ea]
+            if len(ea) == 1:
+                aa = aa[:j] + ea + aa[j + 1 :]
+        return aa
+
+
+# endregion TranslExcept
 
 
 def parse(gff: PathLike, fa: PathLike | None = None):
@@ -339,6 +423,8 @@ def extract(
                 seq: SeqRecord = fet.extract(rec + rec)
             else:
                 seq = fet.extract(rec)
+            seq.features.clear()  # Drop features from `SeqFeature.extract`
+            seq.features.append(fet)
             # endregion extract seq
             if fet_type == "CDS":
                 if len(seq) < min_gene_length:
@@ -348,8 +434,13 @@ def extract(
                 else:
                     seq.annotations["transl_table"] = fet.qualifiers.get(
                         "transl_table", ["Standard"]
-                    )
-                seq.features.append(fet)
+                    )[0]
+                    seq.annotations["frame"] = check_frame(fet.qualifiers)
+                    if "transl_except" in fet.qualifiers:
+                        assert isinstance(fet.location, SimpleLocation)
+                        seq.annotations["transl_except"] = TranslExcept.to_str(
+                            fet.location, *fet.qualifiers["transl_except"]
+                        )
                 seq.annotations["partial"] = fet.qualifiers.get("partial", ["00"])
             seq.id = call_gene_id(rec.id, fet)
             seq.description = " # ".join(
@@ -369,15 +460,36 @@ def extract(
             yield seq
 
 
+def check_frame(annot: dict):
+    frame_ = annot.get("frame", [0])
+    frame_str = frame_[0] if isinstance(frame_, list) else frame_
+    return int(frame_str) % 3 if frame_str and frame_str != "." else 0
+
+
 def translate(rec: SeqRecord, fet: SeqFeature | None = None, auto_fix=True):
-    annotations = fet.qualifiers if fet is not None else rec.annotations
-    frame_str = annotations.get("frame", [0])[0]
-    frame = int(frame_str) % 3 if frame_str and frame_str != "." else 0
-    seq = rec[frame:].translate(table=annotations.get("transl_table", ["Standard"])[0])
-    if auto_fix and annotations.get("partial", [frame])[0] == "0" and seq.seq[0] != "M":
-        # here, if partial is "true" in refseq database, will not fix
-        # elif frame is not 0, will not fix
-        seq.seq = "M" + seq.seq[1:]
+    if fet is None:
+        annotations = rec.annotations
+    else:
+        annotations = {
+            k: fet.qualifiers[k][0]
+            for k in ("transl_table", "frame", "partial")
+            if k in rec.annotations
+        }
+        if "transl_except" in fet.qualifiers:
+            assert isinstance(fet.location, SimpleLocation)
+            annotations["transl_except"] = TranslExcept.to_str(
+                fet.location, *fet.qualifiers["transl_except"]
+            )
+    frame = check_frame(annotations)
+    seq = rec[frame:].translate(table=str(annotations.get("transl_table", "Standard")))
+    seq.annotations.update(annotations)
+    if auto_fix:
+        if annotations.get("partial", str(frame)) == "0" and seq.seq[0] != "M":
+            # here, if partial is "true" in refseq database, will not fix
+            # elif frame is not 0, will not fix
+            seq.seq = "M" + seq.seq[1:]
+        if "transl_except" in annotations:
+            seq.seq = TranslExcept.modify(seq.seq, str(annotations["transl_except"]))
     return seq