+++ tests/__init__.py
@@ -2,7 +2,7 @@
 """
  * @Date: 2023-10-22 20:59:23
  * @LastEditors: hwrn hwrn.aou@sjtu.edu.cn
- * @LastEditTime: 2024-07-02 22:59:42
+ * @LastEditTime: 2025-01-31 12:11:35
  * @FilePath: /genome/tests/__init__.py
  * @Description:
 """
@@ -14,7 +14,7 @@ from tempfile import TemporaryDirectory
 from typing import Callable
 
 
-sys.path.insert(0, str(Path(__file__).parent.parent.parent))
+sys.path.insert(0, str(Path(__file__).parent.parent))
 
 test_temp = Path(__file__).parent / "temp"
 test_files = Path(__file__).parent / "file"
@@ -25,4 +25,5 @@ def temp_output(f: Callable[[Path], None]):
         with TemporaryDirectory(prefix=str(test_temp)) as _td:
             f(Path(_td))
 
+    _f._func = f  # type: ignore[attr-defined]
     return _f