+++ changelog.md
@@ -1,7 +1,7 @@
 <!--
  * @Date: 2023-08-07 15:18:41
  * @LastEditors: hwrn hwrn.aou@sjtu.edu.cn
- * @LastEditTime: 2025-01-23 11:25:53
+ * @LastEditTime: 2025-02-07 11:46:16
  * @FilePath: /genome/changelog.md
  * @Description:
 -->
@@ -44,6 +44,7 @@ changelog for genome
     - muscle in `tree.smk`
     - extract_fna_ko in `genomedb.smk`
     - drep methods
+    - update `gff` and `gene_statistic`  to handle transl_execpt case such as `Sec`
 - 0.2.4:
   - feat:
     - `UniRefClu` method to cluster genes